add isAdmin check to find all and delete all api handlers

+++ src/api/placemark-api.js
@@ -53,6 +53,9 @@ export const placemarkApi = {
     }, 
     handler: async function(request, h) {
       try {
+        if (!request.auth.credentials.isAdmin) {
+          return Boom.forbidden("Only administrators can perform this action");
+        }
         const placemark = await db.placemarkStore.getPlacemarkById(request.params.id);
 
         if (!placemark) {
@@ -75,6 +78,9 @@ export const placemarkApi = {
       strategy: "jwt",
     },
     handler: async function(request,h) {
+      if (!request.auth.credentials.isAdmin) {
+        return Boom.forbidden("Only administrators can perform this action");
+      }
       try {
         await db.placemarkStore.deleteAllPlacemarks();
         return h.response().code(204);
